Created an inner builder class to make constructor private, so that Product objects only can be created true the builder. Code update. Refactored package name.
