+++ src/main/java/Exercise3/repository/ProductRepository.java
@@ -1,0 +1,4 @@
+package exercise3.repository;
+
+public interface ProductRepository {
+}
